# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "PizzaHub API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/cart": {
            "get": {
                "tags": [
                    "Cart"
                ],
                "summary": "Получить корзину пользователя или гостевую корзину",
                "description": "Возвращает содержимое корзины. Если пользователь авторизован, используется его корзина. Если нет, нужно передать `cart-id` в параметрах запроса.",
                "operationId": "getCart",
                "parameters": [
                    {
                        "name": "cart-id",
                        "in": "query",
                        "description": "ID гостевой корзины (если пользователь не авторизован)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "example": "5497a034-39c3-4d14-af0b-675f094926a1"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Корзина успешно получена",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "properties": {
                                                "success": {
                                                    "type": "boolean",
                                                    "example": true
                                                },
                                                "data": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "session_id": {
                                                            "type": "string",
                                                            "example": null,
                                                            "nullable": true
                                                        },
                                                        "user_id": {
                                                            "type": "integer",
                                                            "example": 1,
                                                            "nullable": true
                                                        },
                                                        "items": {
                                                            "type": "array",
                                                            "items": {
                                                                "properties": {
                                                                    "id": {
                                                                        "type": "integer",
                                                                        "example": 57
                                                                    },
                                                                    "product_id": {
                                                                        "type": "integer",
                                                                        "example": 5
                                                                    },
                                                                    "name": {
                                                                        "type": "string",
                                                                        "example": "Кантри-пицца"
                                                                    },
                                                                    "quantity": {
                                                                        "type": "integer",
                                                                        "example": 2
                                                                    },
                                                                    "price": {
                                                                        "type": "string",
                                                                        "example": "438.55"
                                                                    },
                                                                    "stock": {
                                                                        "type": "integer",
                                                                        "example": 2
                                                                    },
                                                                    "dough_type": {
                                                                        "type": "string",
                                                                        "example": "Тонкое"
                                                                    },
                                                                    "total_price": {
                                                                        "type": "number",
                                                                        "format": "float",
                                                                        "example": 877.1
                                                                    }
                                                                },
                                                                "type": "object"
                                                            }
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "total_price": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 877.1
                                                },
                                                "message": {
                                                    "type": "string",
                                                    "example": "Products retrieved successfully"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        {
                                            "properties": {
                                                "success": {
                                                    "type": "boolean",
                                                    "example": true
                                                },
                                                "data": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "integer",
                                                            "example": 33
                                                        },
                                                        "session_id": {
                                                            "type": "string",
                                                            "example": "5497a034-39c3-4d14-af0b-675f094926a1"
                                                        },
                                                        "user_id": {
                                                            "type": "integer",
                                                            "example": null,
                                                            "nullable": true
                                                        },
                                                        "items": {
                                                            "type": "array",
                                                            "items": {
                                                                "type": "object"
                                                            }
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "total_price": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 0
                                                },
                                                "message": {
                                                    "type": "string",
                                                    "example": "Products retrieved successfully"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Корзина не найдена или неверный `cart-id`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "properties": {
                                                "error": {
                                                    "type": "string",
                                                    "example": "Cart not found"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        {
                                            "properties": {
                                                "status": {
                                                    "type": "string",
                                                    "example": "error"
                                                },
                                                "message": {
                                                    "type": "string",
                                                    "example": "Cart not found for session ID 9ba4c140-fc7e-4720-86a5-0aa1fкак"
                                                },
                                                "code": {
                                                    "type": "integer",
                                                    "example": 404
                                                },
                                                "details": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Cart"
                ],
                "summary": "Создать корзину для гостя",
                "description": "Создает новую гостевую корзину. Если пользователь авторизован, вернётся существующая корзина.",
                "operationId": "createCart",
                "responses": {
                    "201": {
                        "description": "Новая корзина успешно создана (только для гостей)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 35
                                                },
                                                "session_id": {
                                                    "type": "string",
                                                    "example": "f88cd52d-d042-47ae-a148-12256eb8e693"
                                                },
                                                "user_id": {
                                                    "type": "integer",
                                                    "example": null,
                                                    "nullable": true
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "total_price": {
                                            "type": "number",
                                            "format": "float",
                                            "example": 0
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "New cart created."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Корзина уже существует (для авторизованных пользователей)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "session_id": {
                                                    "type": "string",
                                                    "example": null,
                                                    "nullable": true
                                                },
                                                "user_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "items": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "example": 57
                                                            },
                                                            "product_id": {
                                                                "type": "integer",
                                                                "example": 5
                                                            },
                                                            "name": {
                                                                "type": "string",
                                                                "example": "Кантри-пицца"
                                                            },
                                                            "quantity": {
                                                                "type": "integer",
                                                                "example": 2
                                                            },
                                                            "price": {
                                                                "type": "string",
                                                                "example": "438.55"
                                                            },
                                                            "stock": {
                                                                "type": "integer",
                                                                "example": 2
                                                            },
                                                            "dough_type": {
                                                                "type": "string",
                                                                "example": "Тонкое"
                                                            },
                                                            "total_price": {
                                                                "type": "number",
                                                                "format": "float",
                                                                "example": 877.1
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "total_price": {
                                            "type": "number",
                                            "format": "float",
                                            "example": 877.1
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Cart already exists."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}